Fix missing c compiler
The GraalVM native image is unable to locate the given bazel c++ compiler
as noted by https://github.com/andyscott/rules_graal/issues/14.

The cause of this is that GraalVM appears to spawn a new project with
a working directory of a new temp file, as a result the path to the
c++ compiler needs to be absolute.

This adds a hacky resolution of the issue by prepending pwd to the
given compiler path before invoking GraalVM.

Additionally, the C compilation variables required by the bazel c wrapper
have been added to the environment variable before execution.